going back and making basic corrections based on test output, tomorrow morning will focus on memfunctions and string functions, then will focus on advanced part 2 string funciont sin the afternoon

+++ ft_strncat.c
@@ -6,7 +6,7 @@
 /*   By: mgould <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/29 15:42:39 by mgould            #+#    #+#             */
-/*   Updated: 2016/11/29 15:46:07 by mgould           ###   ########.fr       */
+/*   Updated: 2016/11/30 17:59:30 by mgould           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char *ft_strncat(char *s1, const char *s2, size_t n)
 {
-	int i;
-	int j;
+	size_t i;
+	size_t j;
 
 	i = 0;
 	j = 0;
